Move select2-multi-suggest to a base component
